refactor: Text Indicator 관심사 분리 (#15)
